+++ spec/brainstem/concerns/presenter_dsl_spec.rb
@@ -21,15 +21,21 @@ require 'brainstem/concerns/presenter_dsl'
 #   field :description, :string
 #   field :updated_at, :datetime
 #   field :dynamic_title, :string, dynamic: lambda { |model| model.title }
-#   field :secret, :string, 'a secret, via secret_info',
+#   field :secret, :string,
+#         info: 'a secret, via secret_info',
 #         via: :secret_info,
 #         if: [:user_is_bob, :title_is_hello]
+#
+#   field :member_ids, :array,
+#         item_type: :integer,
+#         dynamic: lambda { |model| model.members.pluck(:id) }
 #
 #   with_options if: :user_is_bob do
 #     field :bob_title, :string,
 #           info: 'another name for the title, only for Bob',
 #           via: :title
 #   end
+#
 #   fields :nested_permissions do
 #     field :something_title, :string, via: :title
 #     field :random, :number, dynamic: lambda { rand }
@@ -118,7 +124,6 @@ describe Brainstem::Concerns::PresenterDSL do
     end
   end
 
-
   # sort_order :created_at, ".created_at"
   describe 'the sort_order block' do
     let(:value)   { "widgets.created_at" }
@@ -158,7 +163,6 @@ describe Brainstem::Concerns::PresenterDSL do
       end
     end
   end
-
 
   describe 'the conditional block' do
     before do
@@ -218,11 +222,16 @@ describe Brainstem::Concerns::PresenterDSL do
               via: :secret_info,
               if: [:user_is_bob, :title_is_hello]
 
+        field :member_ids, :array,
+              item_type: :integer,
+              dynamic: lambda { |model| model.members.pluck(:id) }
+
         with_options if: :user_is_bob do
           field :bob_title, :string,
                 info: 'another name for the title, only for Bob',
                 via: :title
         end
+
         fields :nested_permissions do
           field :something_title, :string, via: :title
           field :random, :number, dynamic: lambda { rand }
@@ -231,15 +240,24 @@ describe Brainstem::Concerns::PresenterDSL do
     end
 
     it 'is stored in the configuration' do
-      expect(presenter_class.configuration[:fields].keys).to match_array %w[updated_at dynamic_title secret bob_title nested_permissions]
+      expect(presenter_class.configuration[:fields].keys)
+        .to match_array %w[updated_at dynamic_title secret member_ids bob_title nested_permissions]
+
       expect(presenter_class.configuration[:fields][:updated_at].type).to eq :datetime
       expect(presenter_class.configuration[:fields][:updated_at].description).to be_nil
+
       expect(presenter_class.configuration[:fields][:dynamic_title].type).to eq :string
       expect(presenter_class.configuration[:fields][:dynamic_title].description).to be_nil
       expect(presenter_class.configuration[:fields][:dynamic_title].options[:dynamic]).to be_a(Proc)
+
       expect(presenter_class.configuration[:fields][:secret].type).to eq :string
       expect(presenter_class.configuration[:fields][:secret].description).to be_nil
       expect(presenter_class.configuration[:fields][:secret].options).to eq({ via: :secret_info, if: [:user_is_bob, :title_is_hello] })
+
+      expect(presenter_class.configuration[:fields][:member_ids].type).to eq :array
+      expect(presenter_class.configuration[:fields][:member_ids].options[:item_type]).to eq 'integer'
+      expect(presenter_class.configuration[:fields][:member_ids].options[:dynamic]).to be_a(Proc)
+
       expect(presenter_class.configuration[:fields][:bob_title].type).to eq :string
       expect(presenter_class.configuration[:fields][:bob_title].description).to eq 'another name for the title, only for Bob'
       expect(presenter_class.configuration[:fields][:bob_title].options).to eq(
@@ -263,11 +281,17 @@ describe Brainstem::Concerns::PresenterDSL do
           field :updated_at, :datetime, info: 'this time I have a description and condition'
         end
       end
-      expect(presenter_class.configuration[:fields].keys).to match_array %w[updated_at dynamic_title secret bob_title nested_permissions]
-      expect(subclass.configuration[:fields].keys).to match_array %w[updated_at dynamic_title secret bob_title title nested_permissions]
+
+      expect(presenter_class.configuration[:fields].keys).
+        to match_array %w[updated_at dynamic_title secret member_ids bob_title nested_permissions]
+      expect(subclass.configuration[:fields].keys).
+        to match_array %w[updated_at dynamic_title secret member_ids bob_title title nested_permissions]
+
       expect(presenter_class.configuration[:fields][:updated_at].description).to be_nil
       expect(presenter_class.configuration[:fields][:updated_at].options).to eq({})
-      expect(subclass.configuration[:fields][:updated_at].description).to eq 'this time I have a description and condition'
+
+      expect(subclass.configuration[:fields][:updated_at].description).
+        to eq 'this time I have a description and condition'
       expect(subclass.configuration[:fields][:updated_at].options).to eq(
         if: [:some_condition, :some_other_condition],
         info: 'this time I have a description and condition'
@@ -320,8 +344,11 @@ describe Brainstem::Concerns::PresenterDSL do
           field :something, :string, via: :title
         end
       end
-      expect(presenter_class.configuration[:fields].keys).to match_array %w[updated_at dynamic_title secret bob_title nested_permissions]
-      expect(subclass.configuration[:fields].keys).to match_array %w[updated_at dynamic_title secret bob_title nested_permissions new_nested_permissions]
+
+      expect(presenter_class.configuration[:fields].keys)
+        .to match_array %w[updated_at dynamic_title secret member_ids bob_title nested_permissions]
+      expect(subclass.configuration[:fields].keys)
+        .to match_array %w[updated_at dynamic_title secret member_ids bob_title nested_permissions new_nested_permissions]
 
       expect(presenter_class.configuration[:fields][:nested_permissions][:something_title].type).to eq :string
       expect(presenter_class.configuration[:fields][:nested_permissions][:random].type).to eq :number